refactor(isolation): extract screen/docker output parsers into isolation-runner.parsers.lib.mjs
The CI file-headroom check warns for every file above 1350 lines (issue #1593:
long files cause concurrent PR merge conflicts). isolation-runner.lib.mjs was
1357 lines.

The pure output parsers (screen status, session list, exit footer, execution
uuid, docker unknown-exit detection) have no dependency on the runner's I/O, so
they move to their own module and are re-exported unchanged: importers and
tests are unaffected.

isolation-runner.lib.mjs: 1357 -> 1084 lines.

Refs #2175
